despesas-ia/pyproject.toml
[project]
name = "despesas-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[project.scripts]
start = "despesas_ia.api.endpoints:main"
